# Return (or panic) early when a problem is encountered