…he runtime

of *Constraints.AllSatisfied and (IMO) makes Constraint.Satisfied simpler.

Add a benchmark of the Sudoko test to show performance improvement.

Baseline:

$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/gnboorse/centipede
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkSudoku-8            538           2153808 ns/op
PASS
ok      github.com/gnboorse/centipede   2.478s

With this change:

$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/gnboorse/centipede
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkSudoku-8            682           1703352 ns/op
PASS
ok      github.com/gnboorse/centipede   3.491s